// client/src/redux/interfaces/types.ts
export enum EmployeeActionTypes {
  ADD_EMPLOYEE = 'ADD_EMPLOYEE',
  REMOVE_EMPLOYEE = 'REMOVE_EMPLOYEE',
  GET_EMPLOYEE = 'GET_EMPLOYEE',
  UPDATE_EMPLOYEE = 'UPDATE_EMPLOYEE',
  SET_ERRORS = 'SET_ERRORS',

  GETTING_EMPLOYEE = 'GETTING_EMPLOYEE',
  UPDATING_EMPLOYEE = 'UPDATING_EMPLOYEE',
  REMOVING_EMPLOYEE = 'REMOVING_EMPLOYEE',
  ADDING_EMPLOYEE = 'ADDING_EMPLOYEE',
}

export type Action = {
  type: EmployeeActionTypes;
  payload?: any;
};